pyproject.toml
[project]
name = "rhea-mcp"
version = "0.1.1b2"
description = "A scalable MCP (Model Context Protocol) tool framework to serve thousands of biomedical tools for Large Language Models."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "academy-py>=0.2.0",
    "asyncpg>=0.30.0",
    "black>=25.1.0",
    "cachetools>=6.1.0",
    "cheetah3>=3.2.6.post1",
    "conda-pack>=0.8.1",
    "coverage[toml]>=7.9.1",
    "debugpy>=1.8.14",
    "fastapi[standard]>=0.115.14",
    "filetype>=1.2.0",
    "greenlet>=3.2.3",
    "griffe-inherited-docstrings>=1.1.1",
    "kubernetes>=33.1.0",
    "matplotlib>=3.10.5",
    "mcp[cli]>=1.10.1",
    "minio>=7.2.15",
    "mkdocs>=1.6.1",
    "mkdocs-autorefs>=1.4.2",
    "mkdocs-material>=9.6.17",
    "mkdocs-mermaid2-plugin>=1.2.1",
    "mkdocstrings[python]>=0.30.0",
    "mypy>=1.16.1",
    "openai>=1.93.0",
    "parsl>=2025.6.23",
    "pgvector>=0.4.1",
    "prometheus-client>=0.22.1",
    "proxystore[all]>=0.8.2",
    "psycopg2-binary>=2.9.10",
    "pymdown-extensions>=10.16.1",
    "pyright>=1.1.403",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "pytest-cov>=6.2.1",
    "pytest-xdist>=3.8.0",
    "python-magic>=0.4.27",
    "requests>=2.32.4",
    "shellcheck-py>=0.10.0.1",
    "sqlalchemy>=2.0.41",
    "tqdm>=4.67.1",
    "trio>=0.30.0",
    "zstandard>=0.23.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["rhea*"]
